requirements.txt
fastapi
gradio>=5.39.0
gradio_client
Markdown==3.4.1
numpy==1.25.0
requests==2.32.3
sentencepiece==0.1.98
tokenizers==0.13.3
#torch>=2.1.0
#torchvision>=0.16.0
uvicorn==0.21.1
wandb==0.14.0
httpx==0.27.2
transformers==4.28.1
accelerate>=1.1.1
scikit-learn==1.2.2
sentencepiece==0.1.98
einops==0.7.0
einops-exts==0.0.4
timm==0.9.8
openai-clip==1.0.1
fsspec==2024.2.0
kornia==0.6.9
matplotlib==3.7.1
ninja==1.11.1
omegaconf==2.3.0
open-clip-torch==2.17.1
opencv-python==4.7.0.72
pandas==2.0.1
Pillow==9.4.0
pytorch-lightning==2.1.2
PyYAML==6.0
scipy==1.9.2
tqdm==4.65.0
#triton==2.1.0
urllib3==2.2.3
webdataset==0.2.48
xformers>=0.0.20
facexlib==0.3.0
diffusers
accelerate
bitsandbytes
k-diffusion
#mmgp
pydantic==2.10.6
moviepy==2.2.1
